Add YUIDoc build process to the Gruntfile
Note that the docs/ directory is not checked in:
At present, we intend for the docs to live on the gh-pages branch,
where they will to serve as the public face of this module.

We may revisit this decision when we hit 1.0, but for the time being
all documentation updates should be committed to gh-pages.

Which, I might add, does not exist at the time of this commit. TODO!!