Add WP.site static convenience method for creating new sites
The WP.com API client (https://github.com/Automattic/wpcom.js) provides a useful
module.site() method, for constructing a client instance bound to a particular
url. This seems like a convenient shortcut for people coming in from that API, as
well as a nice abstraction for any prospective users who aren't familiar with the
'new' syntax, so I've added .site() as a static method on the WP object.