Move path generation into WPRequest and change method names
PostsRequest URI generation is now handled within WPRequest, which
builds a path based on the _path.template string on the request
instance. Definition of the path structure and validators should
happen on the endpoint handler level; putting those together into a
final URL should be generic (and therefore live in WPRequest).

Identifiers changed:

* ._action, ._id, and ._actionId are now grouped in ._path.values
* .generateRequestUri() becomes ._renderURI()
* ._pathStr() becomes ._renderPath