Add a .root() method for querying custom endpoints
WooCommerce, and soon other plugins, will be adding custom endpoints to
the API. This method is only a stopgap, but .root() provides a way to
specify a request against one of these custom API endpoints by providing
a root-relative path to the new endpoint.

Note that no path interpolation will occur, so the provided path must be
a dumb string, not a custom(/:url)(/:template) of any sort!