Specify explicit indices for query parameter array syntax (fixes #89)
superagent was transforming the `type[]=type1&type[]=type2` portion of
the URL into `type[][0]=type1&type[][1]=type2` because it detected
`type[]` as a repeating parameter key and added its own array brackets.

This syntax is invalid, and results in an error on the WordPress side.

By removing the prepareParameters method we could eliminate all of our
custom handling for array syntax, and trust that superagent will take care
of it: however, this introduces an implicit dependency on superagent's
continuing to use a query parsing library that causes this behavior.

A more robust solution is to use the same "qs" module to parse our own
query strings that superagent uses internally. Using it here not only
simplifies our code (replacing two methods, `prepareFilters` and
`generateQueryString`), but also removes our dependency on superagent's
internal behavior: instead of planning for the request library to
transform our URI in a particular way, format our URI in that manner to
begin with so that no further transformation occurs.