Move all test specs into tests/unit/ folder
This sets us up to have an integration test suite that we can run
against a live demo site, or against replay-stored data files.

Eventually use of replay and the integration suite should supplant
the use of the test cases we currently have which only assert URLs are
generated correctly.