Alter methods to use new v2 top-level parameters
Query strings can be made more concise by utilizing non-filter query
parameters where appropriate. (`?author=` is shorter than
`filter[author]=`, for example, and accomplishes the same thing).
Brevity is the soul of wit, but it's also a great way to avoid needing
to toss a lot of other to-be-HTML-encoded brackets into the query.