Add basic integration tests for tag/category filtering
Note that specific categories were chosen to make the multi-cat query
work propertly:

The issue with using e.g. "Media" was that the category_name is "media_2",
so using just "Media" in the query didn't find a hit. Markup is unique in
our sample data (it's not a tag as well), so it is a reliable category to
which we can add "pustule" to filter down to a single post.