Adjust WPRequest's .auth() to permit specifying user/pass as an object
It was inconsistent that authentication could be specified as key-value
pairs when creating the WP instance, but not when calling the .auth()
method on a WPRequest chain. This alters WPRequest's .auth() method to
take EITHER the previously-supported two-string-arguments for username
and password, OR an object with 'username' and 'password' strig properties.