Remove .version() method; use .namespace() method to rule them all
See #141: in the API code, there is no conception of a "namespace"
separate from a "version." That is, `wp/v2` itself as a whole is the
"namespace." In light of this, providing separate ns & v methods,
and treating them as separate strings, would make it more difficult
to support endpoints that do not confirm to what amounts to an arbitrary
naming standard.

This PR removes the `.version()` method, and uses `.namespace()` to
set a combined string of format "wp/v2" for the v2 beta API defaults.