Add initial support for querying custom API endpoints
This begins to address #140, by implementing an .endpoint method
that will create and return a factory method which can create
CollectionRequest instances against the provided endpoint.

Downsides to this approach:

- Somewhat duplicative w/r/t .root()
- Presumes that only resource(/:id) format collections will be used
  - Actually supporting an arbitrary structure of URL will eventually
    be necessary in order to support discovery-based generation; however,
    at present we haven't worked out a solid approach for parsing PCRE
    named groups.

Upsides:

- Provides a vector to reintroduce .registerType very cleanly

The current plan is to merge this, get feedback, and iterate prior to
making a formal decision on supporting this or using an alternative
approach prior to releasing v0.7.