Move filter & parameter chaining methods into mixins
This PR extracts the filtering and query parameter methods into mixin
functions that can be added to a constructor to grant it the specified
methods. In addition to this significant (architectural) change, filters
and parameter setters no longer have implicit interdependencies: with
the exception of .param itself, which is a part of the base WP-Request
class, all filter methods depend on each other directly where needed so
that any filter can be included into a subclass with no restrictions.

The goal of this patch is to

1. decrease code duplication between endpoints
2. better surface and test edge cases in parameter and filter methods
3. simplify the ability to add filtering functionality to programatically-
   generated endpoint handlers (see #145)

To accomplish these goals I

- moved CollectionRequest filter methods into a filter mixin module
- moved CollectionRequest parameter methods into a parameter mixin module
- moved filter and parameter mixins from other endpoints (page, taxonomies,
  comments) into the filter and parameter mixin modules as appropriate
- unit tested the new filter and parameter mixin modules
- adjusted how `.filter` and `.param` function to address edge cases
  exposed during testing, such as unexpected behavior when calling
  `.filter` or `.param` with no second argument.
- allowed filters to initialize their own private state objects (the
  `this._filters` and `this._taxonomyFilters` objects), so that they
  can be included into a prototype without any further pre-work required
  to make them function with that constructor
- removed all existing explicit definition of `this._filters` and
  `this._taxonomyFilters` objects in endpoint request handler constructors,
  since those objects will now be created only when needed

There is currently some duplicate testing logic between the mixin files
and the request handler tests themselves; this is not seen as a detriment
but since the mixins are fully tested the assertions in the constructor
test suites may eventually be replaced with tests that verify the mixins
have been applied properly.