Move Testing section to top of Contributors.md
Testing is a key part of onboarding to the library, much more important
than code style conformity: it should be the first thing that a user
sees when they open the contributing guide. (Leading with code style
just feels pedantic to me now.)

We also should make it clear that tests are useful, but that we can work
with a feature submitter to get them in -- and that a test demonstrating
a bug can be just as useful as a bug patch, if not more so. (because we
will need to write that test eventually anyway).

This also makes very minor line edits to the patch from #160.

Closes #160.