Switch .post() and .put() to .create() and .update()
This one has been a long time coming: POSTing a post to the posts endpoint
is a linguistic nightmare, and HTTP verbs are part of the API interface
that is best kept away from the users.

This alters all of the existing GET, POST, PUT, DELETE and HEAD methods
to be (semi)private and prefixed by `_http`, e.g. `_httpDelete`: The
public interfaces for these (`.get`, `.create`, `.update`, `.delete`
and `.headers`).

`.post` and `.put` are maintained for the time being for back-compat
but will be removed in 1.0.