Remove dependency on full Lodash package
Lodash does a lot of things we do not especially need in this library. It
is more straightforward to only consume those lodash modules we need (and
from a file-size perspective this may also assist in the browser-side usage
of the library via a bundler such as Browserify or Webpack).

The primary lodash methods in use were installed as discrete dependencies:

- Depend on `lodash.reduce` explicitly for use in reducing over objects
- Depend on `lodash.uniq` explicitly for array de-duplication
- Depend on `lodash.union` explicitly for computing array union

Usage of Lodash as a general utility was eliminated entirely wherever it
was convenient, including shifting to:

- Use `Array.isArray` over of `_.isArray`
- Use `typeof` over `_.isString` and `_.isNumber`
- Use `=== null` over `_.isNull`
- Create first-party `keyValToObj` utility to replace `_.zipObject`,
  which has a confusing API and which seems to have been renamed to
  `fromPairs` in a recent lodash update

Lodash-independent changes:

- Use `(=!)== undefined` instead of `typeof undefined`, given strict mode

Removing the full lodash saves ~25kb in the Webpacked bundle.