Switch from Bluebird to es6-promise polyfill
We are not using any special functionality from Bluebird (see
http://bluebirdjs.com/docs/getting-started.html), and switching to the
lightweight es6-promise polyfill saves a decent number of kilobytes
when creating a browser bundle.

Using the core build of Bluebird was considered, but this library's
Promise needs are not so complex that this option seemed preferable to
simply using a spec-compliant polyfill.