Do not log a warning when skipping preexisting methods during discovery
As setters are inferred from the shape of the API routes, there are
situations where the names parts of a route will result in setter fn
names that conflict with (i.e. are the same as) preexisting methods
on the WPRequest object. The `.post` query terminator that triggers
a POST request is a good example, as the part of the route indicating
post ID for e.g. a revision is named as "post".

In this situation we skip assigning that setter, as in most cases it will
not preclude accessing that resource through other means (the .id method
will set the post ID in this case, and the revisions method will set the
revision ID). Historically, however, we did log a warning when this
occurred, to note that certain methods may not be available as expected.

This warning is distracting and causes concern that the library is not
functioning as intended. This PR removes the warning, silently skipping
any pre-existing methods when assigning setters. This suits the majority
use-case without preventing edge-case users from taking advantage of the
registerRoute method to craft a custom handler for an arbitrary endpoint.

Closes #199