Do not unnecessarily lowercase path part setter names
The logic used to create setter names for path parts within a resource
handler incorrectly assumed that all path part components should be
lowercased before being converted from snake_case to camelCase. If a
resource already employed a camelCase component (and it is perfectly
valid to do so), then the setter for thatComponent would be created as
`thatcomponent`, not `thatComponent`. This is unintuitive behavior (and
actually raises the question of whether the token should be mutated at
all), so this PR removes the offending `toLowerCase` method and documents
the naming conversion logic in the readme.

Fixes #198, props @joneslloyd for the bug report