Release v0.9.0 "Since I Left You"
v0.9 brings two long-neglected features to the foreground: media
support and browser-side usage. This release is named in homage to
The Avalanches' absurdly brilliant _Since I Left You_.

**Media Handling**

`wp.media()` requests now provide a `.file()` chaining method, which can
be used to specify the file system path of a file to upload (in node) or
the contents of a file input (in a browser). This permits images and
other media to be uploaded to the WordPress site in the same command
which creates the media record.

**Browserify/Webpack Support**

Requiring `wpapi` as a dependency in a Browserified or Webpacked build
now results in a much slimmer file bundle,

**Miscellaneous Fixes**

- Irritating warning messages generated on WP instantiation have been
  removed (reported by @elyobo)
- registerRoute no longer improperly lowercases camelCased path part
  setters (reported by @joneslloyd)
- The WP constructor now throws a nice error if the endpoint property
  is not a string, not just if it is falsy (props @sdgluck)
- JSDoc fix (props @sdgluck)

**Props**

For submitting pull requests and patches:
@bt
@sdgluck

For opening issues and asking or discussing questions:
@elyobo
@joneslloyd
@mrkrumhausen
@satish9323
@smedegaard
@tommedema
@vtripolitakis
@z-avanes