Add a Jekyll-based documentation site
There's a wealth of content in the readme, and an opportunity to write
additional user guides. However, the documentation site currently features
only the generated API docs, and ever since #145 the generated docs are
not as robust as they were before. To better highlight the documentation
contained within the REAMDE, in this PR we

- Add grunt task to zip built browser bundles (for #196)
- Implement a task to generate jekyll-ready pages from README
- Generate a documentation index with deep links to the subheds of
  content extracted from the README file
- Update CONTRIBUTING and README files to generate better output
- Add a script to print out tasks necessary to release docs branch
- Add a docs site favicon and add IE fallback favicon link
- Add a convenience package script to run jekyll in dev mode
- Add a grunt task to clean leftover files after docs release
- Reorganize Grunt tasks into modules within `build/grunt/`
- Lint build and bin directories, with support for es2015 syntax
- Support arbitrary host when running jekyll in local dev environment

There's a lot that could be done to improve the generated documentation, but as-is this closes #196

Closes #205 (this is a squashed PR)