Allow authentication credentials to be set after autodiscovery
As noted in #216, the process of authenticating after autodiscovery is
not obvious: it is technically possible, but the manner documented uses

```js
WP.discover( 'http://my-site.com' ).then(function( site ) {
  // Call .auth method on individual query
  return site.users().auth( credentials ).me();
}).then(function( user ) { /* ... */ });
```

which is verbose and not obvious in the way that instantiating a WP
instance with authentication "baked in" can be.

To resolve the issue, this PR exposes & documents the (existing) `.auth`
method on the WP site instances themselves: this method is identical to
WPRequest's `.auth`, but sets the provided options on the WP site client
instance itself so that they will apply to _all_ subsequently-generated
API requests.

The easiest way to use authentication with the new system becomes:

```js
WP.discover( 'http://my-site.com' ).then(function( site ) {
  site.auth( credentials );
  // site is now authenticated as normal
});
```

This approach will work with both currently-supported forms of
authentication, Basic HTTP Auth & Cookie/Nonce Auth.

Closes #216