Centralize integration helper methods & improve auth handling
Following #217 and discussion with @gnarf I am updating most tests which
make use of authentication to use two separate WP instances, one that is
auth'd and one that is not. This (to me) improves the readability of the
test suite.

Additional whitespace commit to reformat indentation on integration tests
to follow in this same PR.