Promote _renderURI to a public api & rename to toString
`._renderURI` is arguably the most important single method this library
provides; it would be possible (and should eventually be easy) to use
the query builder interface independently of any specific HTTP library
like superagent, _e.g._ if you wanted to use jQuery or `fetch` to send
requests, so long as you have the ability to get at the full, rendered
URI for the API query you need to execute.

As such, having this be exclusively "private" (or as private as `_methods`
ever can be) functionality has proved confusing. Exposing `_renderURI`
as `.toString` makes it a first-class public citizen of the library,
and makes it more obvious how the query methods lead to the generation
of a URI that is then converted into an HTTP request action by one of
the HTTP terminator methods.

`_requestURI` is now deprecated and will be removed in the next release
(along with the deprecated "http verb" versions of the request terminators)