Extract HTTP behavior from WPRequest into separate transport module
WPRequest currently does two things: It builds a string, then it allows
HTTP requests to be submitted against the resource that URI string
represents. The latter part of that behavior will benefit from being
pluggable, as described in #211 and #213.

Exposing WPRequest itself is something I would like to avoid to keep
the core workings of the library (its URI-builder behavior) safe from
unexpected modification; but exposing the HTTP behavior and enabling
it to be customized or extended not only solves the issue expressed
in the linked issues around cache implementation, but also provides
a better interface for specifying bespoke authentication behavior.

To that end, this commit moves WPRequest's HTTP logic into a separate
http-transport.js module, and injects a _copy_ of that module's
exports into `new WP` objects. The next step will be to make `transport`
a property which may be passed in, and to expose the default transport
on the `WP` object.