Release v0.10.0 "Abaporu"
v0.10 standardizes the name of the constructor exposed by this module
as "WPAPI", and creates a seam through which HTTP interactions can be
customized or overridden. This release is named for Brazilian DJ
Gui Boratto's fourth studio album _Abaporu_ (Kompakt, 2014).

**Module Naming**

The constructor exposed by this module should be referred to as
`WPAPI` in all cases. This has been updated across the documentation
site, source code, and examples within code comments.

There is remaining inconsistency throughout the docs & test suite
between whether an _instance_ of `WPAPI` is referred to as `wp` or as
`site`; this will not be addressed at this time, the identifier used
to hold the WP site client instance is left up to the library consumer.

**Custom HTTP Transport Behavior**

Custom HTTP transport methods may be provided to inject or short-circuit
HTTP behavior for each type of request; methods may be specified via a
`.transport` property on the configuration object passed to the `WP`
constructor, or by passing an object of transport methods to the
`.transport()` method on the instantiated `WP` object.

**Props**

@edygar