Update taxonomies integration test to be forward-compatible
I am modifying wpapi-vagrant-varietal, the VM environment against
which these integration tests are run, to include some custom
post types and taxonomies. This change will prevent that commit
from breaking these integration tests because it switches to
validating that what we expect to be there is there, not that it
is the only content available.