Always use authentication when credentials are available
Up to now, if a username and password (or nonce) were supplied when
instantiating a WPAPI instance, those credentials would not be used
by default. This meant that these requests would be handled differently
in terms of authentication, despite having all the same information:

```js
// Authenticated
site = new WPAPI({
  endpoint: 'some.website.com/wp-json'
}).auth({
  username: 'bananayoshimoto',
  password: 'kitchen'
});
site.posts()...

// NOT Authenticated
site = new WPAPI({
  endpoint: 'some.website.com/wp-json',
  username: 'bananayoshimoto',
  password: 'kitchen'
});
site.posts()...
```

The consumer of the API must have called `.auth()` at some point on the
parent WPAPI object or the instantiated request chain in order for
authentication to be used on the request.

This is unintuitive and inconsistent, and even as the library author
it has been tripping me up. This PR standardizes the behavior so that
authentication will always be used if credentials are available.

Pork barrel:

- Standardize on `site` for the identifier for a WPAPI instance in the
  wpapi mocha test suite
- Update two tests to use the credentials object `.auth()` syntax, not
  the now-deprecated `.auth( 'username', 'password' )` syntax