Breaking: Remove deprecated `.auth( 'user', 'pass' )` method signature
The string username & password signature for the `.auth()` method (both
for WPRequest-derived handlers & WPAPI instances itself) has been
deprecated for the last release or so because it offers less flexibility
vs explicit parameter naming. Were we to add, say, OAuth support, we
would want to be able to specify a token and secret value, but would
not want them to be confusable with the less-recommended Basic Auth
scheme. Removing the signature that uses unnamed strings disambiguates
in a future-oriented way.