Release v0.12.0 "You Only Live Once" (BREAKING CHANGES)
This release makes breaking changes to how endpoint filtering methods
are bound, specifically reacting to the removal of the `?filter` query
parameter which is not present in WordPress 4.7's API endpoints.

New filtering methods:

- `.categories()`
- `.tags()`

Both of these methods support **querying by ID only**. If a slug is
provided it will **not work** in 4.7, due to the removal of `filter`.

**Deprecated** filtering methods:

- `.category()` (for filtering by associated terms; use `.tags()`)
- `.tag()` (for filtering by associated terms; use `.tags()`)

_Note:_ these method names may be retained, but the functionality which
delegates to `.filter()` to query by term slug will be removed.

**Removed** filtering methods:

- `.name()`: use `.slug()` (deprecated previously)
- `.filter()`: available as a mixin but no longer enabled by default
- `.path()` (for Pages)

Props @joaojeronimo, @Ohar, @ryelle

This release is named for w.hatano's latest EP [_You Only Live Once_][yt]

[yt]: https://www.youtube.com/watch?v=V7mnNdkhciQ