Implement basic objectReduce() function to replace lodash.reduce
As identified by webpack-bundle-analyzer, lodash.reduce is the largest
single dependency within the wpapi library (390.08kb overall):

![image](https://cloud.githubusercontent.com/assets/442115/20847911/fe26386e-b89d-11e6-8f43-178d1c944a6e.png)

Lodash.reduce is powerful, but has more code and optimizations than are
needed for our specific use-case of reducing over an object of key-value
pairs in the manner of Array#reduce.

This PR replaces Object.reduce with an API-identical method objectReduce,
which handles the simple case of iterating over basic objects containing
only own-property key-value pairs. This functionality is sufficient for
the needs of node-wpapi, as Array#reduce itself is used for reduction
over arrays and we are generally iterating over basic route definition
JS objects, not complex instance objects.

![bundle analysis after removing lodash.reduce](https://cloud.githubusercontent.com/assets/442115/20848132/c1823c5e-b89e-11e6-8e50-540c42e22b21.png)

After this change our build drops from 421kb to 355kb unminified, and
from 99.4kb to 87.2kb minified.

A unit test suite is added to validate that objectReduce results in
identical output to lodash.reduce for the specified use-case.