Make requests using HTTP transport during auto-discovery process
In #181 auto-discovery was added that permits a consumer of this library
to automatically locate the API root for a given WordPress site (specified
by URL), and in most cases to automatically detect & initialize specific
route handlers for the API routes available on that site.

However, these methods were never updated to account for #221, which
added the ability to replace the default HTTP transport (superagent)
with a custom implementation or a wrapper. This meant that even if
the consumer had specified their own transport, for example jQuery,
there would still be a dependency on superagent should they wish to
use the auto-discovery capabilities of the library.

This PR updates the auto-discovery logic to use the same HTTP transport
layer as the rest of the client, which has the added benefit of removing
a number of duplicative logic that existed in the auto-discovery module.