Add `.status()` query method & fix mixin application bug
This PR adds a `.status()` query method for setting one or more statuses
to return for endpoints that differentiate based on status, such as Posts
(fixes #264). It also allows the `.password` and `.sticky` mixins to be
properly applied; they were not previously auto-set when working with
a collection that supports those parameters as they were not included in
the mixin index.

All three parameters are now tested consistently, & all are documented in
the README file.