Breaking: Remove "merge" argument from `.param` method (Closes #227)
The third "merge" argument in `.param( key, val, merge )` has been on
the chopping block for two months and is now being removed prior to the
1.0 release.

WPRequest instances are intended to be transient: It is much easier
to understand building a single array of values, and then call
`.param( 'key', valuesArr )` once at the end, than it is to mentally
keep track of calling `.param( 'key', value, true )` multiple times.

It is also one of the only mysterious boolean arguments in the entire
library, and we normally aim to avoid those.

It is also only used in one place (the `.parent` mixin), and is the only
place the `lodash.uniq` NPM dependency is used within node-wpapi.
Removing the argument allows us to simplify code in several places and
permits the removal of that dependency library, slightly decreasing the
size of the built browser UMD bundles.