Further slim down default routes JSON bootstrapping file
The routes object is the only part of the default data file that gets
consumed by the API client during bootstrapping, and not all properties
within that dictionary are even used: for example the "required" status
of arguments is not used anywhere in the library. In the spirit of YAGNI
these extraneous values are being removed in preparation for v1.0.