Update README for 4.7, re-point links to reference developer.w.org
Fixes #293

The documentation in the README (and therefore the node-wpapi.wp-api.org
site) is out-of-date: it points to the legacy wp-api.org instead of
the new developer.wordpress.org/rest-api handbook, and it references
plugin-era methods instead of the new, standardized core methods for
cookie authentication.

Props @bkirby989 for reporting this discrepancy.