Handle case where JSON response is returned as content-type text/html
As noted in #146, the API response is not properly parsed by superagent
when a caching plugin is enabled: the `.body` property of the response
ends up null, so an error is thrown when the pagination object is tacked
on. This seems to originate from the caching plugin (WP SuperCache)
forcing a "text/html" content-type header on the API responses.

The workaround in this PR is to detect for the text/html content type,
and to parse the JSON in the response `.text` property should the
response body itself end up empty.

Fixes #146