Release v1.1.0 "Where's Your Head At"
This release introduces a `.setHeaders()` method on both WPAPI objects
and derivatives of WPRequest, to permit arbitrary custom headers to be
dispatched along with an API request.

Inline code documentation has been overhauled and reviewed for
completeness, as well as converted from YUIDoc to JSDoc.

props to @anagio, @andreasvirkus, @entr, @gnarf, @Matthewnie, @mnivoliez,
and @mzalewski for their contributions!

This release is named for Basement Jaxx' 2001 single "Where's Your Head At"