Maintain authentication when making paginated .prev/.next requests
When we create the WPRequest object to handle the previous or next pages
of a collection (when available), we were passing only the endpoint and
the HTTP Transport object -- this creates a valid WPRequest object, but
authentication parameters are then not passed on to the new request
objects. This PR updates the logic within the transport to maintain any
set options on the request, overwriting only the endpoint URI, so that
authentication will be maintained when paging through a collection.

Fixes #325

Props @motleydev