Fixing Integration tests - wrong endpoint name
The use of the [wpapi-vagrant-varietal](https://github.com/kadamwhite/wpapi-vagrant-varietal) vagrant machine, exposes WP API functionality through the ```wpapi.local``` address.  Many of these are using the address ```wpapi.loc``` therefore have replaced these with ```wpapi.local```.  An alternative might be some form of redirection from wpapi.loc to wpapi.local - however I think there are less 'moving parts' by just replacing the names of the servers.  A future improvement might be to move this URL to some form of configuration.

As it currently stands this causes many of the integration tests to fail.  

In addition the credentials stored in the [constants.js](https://github.com/WP-API/node-wpapi/blob/master/tests/integration/helpers/constants.js) do not match any users in the system. These have been set to ```admin```/```password```.

Have also noticed that the unit testing for autodiscovery still uses this server - Not sure that this would still be a unit test - however in order to make the tests pass, have adjusted this accordingly.  Have also raised an issue relating to this : #393