Refactor transport binding to prevent cross-contamination
Because of how transports were previously all bound using the single shared
WPAPI constructor, all tests which purported to set the transport only for
one transport-specific constructor in fact all shared whichever transport
was set last. To work around this, the bindTransport method has been
enhanced to subclass WPAPI, and the definition of the discover method has
therefore moved into that function, rather than the base wpapi.js

.site() is useful regardless of whether a transport is present, so it
remains in the base class; however, we override .site() with a version
which provides instances of the new transport-specific subclasses when
creating a new version of the constructor in bindTransport().

As a plus, this fixes the issues we had in the transport binding tests!
Turns out that wasn't a false negative after all.